%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  print.SmoothSplineLikelihood.R
% 
% on Sun Apr 27 23:25:36 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{print.SmoothSplineLikelihood}
\alias{print.SmoothSplineLikelihood}
\alias{SmoothSplineLikelihood.print}
\alias{print.SmoothSplineLikelihood}
\alias{print,SmoothSplineLikelihood-method}


\title{Prints an SmoothSplineLikelihood object}

\description{
 Prints an SmoothSplineLikelihood object.  A SmoothSplineLikelihood object is returned by
 \code{\link{likelihood.smooth.spline}()}.
}

\usage{\method{print}{SmoothSplineLikelihood}(x, digits=getOption("digits"), ...)}

\arguments{
  \item{x}{Object to be printed.}
  \item{digits}{Minimal number of significant digits to print.}
  \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{internal}
\keyword{methods}
